train_celeba: allow custom dataset for classifier and vae
Added options to allow swapping out the celeba dataset with
any other fuel compatible dataset with 64x64 features. If
the dataset is grayscale instead of color, the color-convert
option can also be used to dynamically transform the
data into color as it comes in.

Note that for train_celeba_classifier the custom fuel dataset
must also have compatible targets, which is unlikely already
the case for common datasets but it certainly possible with
some customization. However, for train_celeba_vae any fuel
dataset with 64x64 features can be used - either with
regularize left off or by using an existing celeba classifier
or any other classifier trained on data close enough to
the target dataset.