train_celeba_classifier: added "allowed" label filter, cli opts
The training can now be done on a subset of the 40 available
CelebA labels. For example, to train only on "lipstick" and
"big lips", just provide the command line option with numbers
that match those columns:

  --allow 6,36

And all other labels will be zeroed out by a Transformer.

Also added the ability to specify other options on the command line:

  --classifier for output file, defaults to "celeba_classifier.zip"
  --batch-size for size of all mini-batches. defaults to 100