train_vae: added --oldmodel option for starting from previous model
Added an option to start training from a prevoius model
checkpoint. Note that this also involved a change to MainLoop
to add the model, which sadly means that previously saved
checkpoints are not usable.

Also updated train_monitoring options to set
before_first_epoch=True as this is useful to verify
that the checkpoint was loaded successfully.