train_vae: deepened network for larger input images
Added more layers to VAE because now running on 256x256
input images. Ultimately, the depth of the network should
be dependent on the input image, but this will do for now.

Also increased recursion limit to get serializatoin working
again and - as a debug tactic - have serialization happening
now before the first epoch.